upgrade Groovy to 5.0.4 with compatibility fixes
Work around GROOVY-11289 union type issue in instanceof checks,
add explicit casts for stricter type coercion in Groovy 5.

+++ src/main/groovy/cz/siret/prank/program/routines/results/EvalResults.groovy
@@ -111,11 +111,11 @@ class EvalResults extends ResultsBase {
     }
 
     int getAvgTrainPositives() {
-        (double)train_positives / runs
+        (int) ((double)train_positives / runs)
     }
 
     int getAvgTrainNegatives() {
-        (double)train_negatives / runs
+        (int) ((double)train_negatives / runs)
     }
 
     double getTrainPositivesRatio() {
